requirements.txt
torch
numpy
pandas
scikit-learn
tqdm
colorama
biopython
fair-esm
huggingface_hub
